[PORT FROM MAIN] Recovery after lost read messaged from pn544
BZ: 21462

Every time phone's screen is toggled, the polling loop of the NFC controller
is modified to enable/disable target discovery for reader and P2P mode.

Occasionally it happens that when disable target discovery sequence is
started the chip seems to be in some transient state: it's transitioning
from the active to the standby mode. It's able to accept the data
on the I2C level, and probably to process it but never emits proper
answer to the host (probably goes to standby instead).
Then, when the host repeats the write after LLC guard timer elapses
in order to retrieve the lost answer from the chip,
the chip wakes up from the standby and processes this request but
treats it as the redundant one (already answered) and requests
the host to continue with the sequence. The host however never got
the answer to the initial request and both sides get out of sync.

To recover from this situation we added to the sequence send of
a RSET frame to resynchronize both sides when they are out of sync.

Change-Id: I29130f8a6d8a77ad552f5e2810ca33c00f077609
Orig-Change-Id: Id66f7a2a75c63c46296a2f047f51b2670b96b2c4
Signed-off-by: Sebastian Niciarz <sebastian.niciarz@intel.com>
Signed-off-by: Ken Wahid <kenx.wahid@intel.com>
Reviewed-on: http://android.intel.com:8080/35419
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/58852
Reviewed-by: Rebraca, DejanX <dejanx.rebraca@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/78403
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>