[PORT FROM MAIN]Fix a race condition in the phOsalNfc deferred timer callback implementation.
BZ: 20885

In phOsalNfc_Timer.c: When a deferred timer callback function (of type
PH_LLCNFC_ACKTIMER) was scheduled to run *before* the previous deferred callback
had completed, a global pointer was overwritten, causing the deferred callback
function to free a yet unprocessed piece of memory, which eventually lead to a
SIGSEGV exception when trying to execute a function pointer stored inside this
freed piece of memory.

A semaphore scheme was added to fix this problem. Now ensuring that a pending
deferred timer callback function gets to complete before a new one is scheduled.

By adding this semaphore scheme, we potentially introduce a delay in the NFC
timer processing. However this delay is introduced only in the rare occasions
that would previously lead to the race condition and crash. And no side-effects
of this delay were observed during 12+ hours of testing on multiple devices.

A future improvement could be to put in place an array- or linked list scheme
instead of the use of a semaphore. I.e. to queue up the deferred callback
events on a list for sequential processing.

Change-Id: I1e3da340ef0c73f7fcbffee7599712909247705d
Signed-off-by: Ken Wahid <kenx.wahid@intel.com>
Reviewed-on: http://android.intel.com:8080/35254
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Reviewed-by: Rebraca, DejanX <dejanx.rebraca@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Sebastian Niciarz <sebastian.niciarz@intel.com>
Reviewed-on: http://android.intel.com:8080/58851
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/75762
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>