LLCP: Accept/Connect socket race condition fix
BZ: 56636

Accept and connect APIs from NDEF client/server java threads are using
the same response callback:
phFriNfc_LlcpTransport_ConnectionOriented_SendLlcp_CB.
In case a connecting is in progress, and at the same time if it happens
that we are receiving something from the other target, the accept method
will fail and return, but it'll leave wrong socket state (eSocket_State)
which will be misinterpreted in connect callback
(phFriNfc_LlcpTransport_ConnectionOriented_SendLlcp_CB). That's why the
crash appears. In order to avoid this situation, the socket state must be
preserved in case of failing accept.

Change-Id: I644477c2288ad4b146c327ac1dd4e095cf144c1f
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/66777
Reviewed-by: Wahid, KenX <kenx.wahid@intel.com>
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/75765
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>