LLC: Fix race condition when sending ACK and I frame at the same time
BZ: 62646

I-frame from host has to be sent in the interval of ACK timeout (1ms).
In case the I-frame is successfully sent, the ACK frame is canceled. If
we have a corner case where the frame is about to be sent, whereas at the
same time the acknowledge time expires (we are at the border of ACK time
interval), then the ACK frame and I frame will be preparing for sending
accessing at the same time the common frame information resources. The
mutex protection has been added in order to avoid this bad scenario.

Change-Id: I1bbb803a4bcf9b09995c23366b4657dd974f9072
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/71595
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Tested-by: lab_cactus <lab_cactus@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/78342
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>