Adding feature for sanity checking the NFC device at init
BZ: 63164

CAN BE UPSTREAMED

Added a feature for sanity checking the NFC device and
communication link (i2c/uart) during DAL initialization.
An error code is returned to upper layers in case of failure.
Improving the error detection and disabling of NFC functions
on boards with faulty NFC hardware.

The enabling of this feature is controlled by the
configuration macro DEVICE_SANITY_CHECK_AT_INIT. The feature
is default disabled.

Change-Id: Idea1e5da9fccbe864591faf256010226ad227ab7
Signed-off-by: Ken Wahid <kenx.wahid@intel.com>
Reviewed-on: http://android.intel.com:8080/73433
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/77423
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>