Add write protection mutex in write response callback
BZ: 91279

Write protection mutex is added in write response callback in order to
ensure that LLC send function completes properly without preemption of
the callback thread. This is necessary to have, because otherwise, in
case of preemption, the frame info structure will not be fully updated
or guard timer could start after receiving a response to I frame.

Change-Id: I9aa4d861038e77cda28c1d7d34d44dea35709622
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/95816
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>