[PORT FROM R42S] Add timeout on i2C read during deinitialization
BZ: 96770

During SW stack deinitialization all timers will be stopped, thus we can
have a situation where we are blocked in read waiting for a chip
response which doesn't come for any reason. Since the timers
are stopped, we will not have any "guard" to monitor this situation.
That's why the timeout has been added in select function of i2c_read to
unblock the RX thread in this case.

Change-Id: Iad649ed54d20cd618a7bf60a387b9acb704c241b
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Signed-off-by: Sebastian Niciarz <sebastian.niciarz@intel.com>
Reviewed-on: http://android.intel.com:8080/101902
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>